// src/composables/games/wil/enums/image.ts
export enum WIL_IMAGE_ID {
  CHARACTER_HERO = "CHARACTER_HERO",
  CHARACTER_HOLY_KNIGHTS_LEADER = "CHARACTER_HOLY_KNIGHTS_LEADER",
  CHARACTER_HOLY_KNIGHTS_SOLDIER = "CHARACTER_HOLY_KNIGHTS_SOLDIER",
  CHARACTER_HOLY_KNIGHTS_MAGICIAN = "CHARACTER_HOLY_KNIGHTS_MAGICIAN",
  DARK_MONSTER_SHADOW = "DARK_MONSTER_SHADOW",

  MINI_CHARACTER_HERO = "MINI_CHARACTER_HERO",
  MINI_CHARACTER_HOLY_KNIGHTS_LEADER = "MINI_CHARACTER_HOLY_KNIGHTS_LEADER",
  MINI_CHARACTER_HOLY_KNIGHTS_SOLDIER = "MINI_CHARACTER_HOLY_KNIGHTS_SOLDIER",
  MINI_CHARACTER_HOLY_KNIGHTS_MAGICIAN = "MINI_CHARACTER_HOLY_KNIGHTS_MAGICIAN",
  MINI_DARK_MONSTER_SHADOW = "MINI_DARK_MONSTER_SHADOW",
  CHARACTER_OTHER = "CHARACTER_OTHER",

  SKILL_SLASH = "SKILL_SLASH",

  TRAINING_ATTACK = "TRAINING_ATTACK",
  TRAINING_DEFENSE = "TRAINING_DEFENSE",
  TRAINING_MIGRATION = "TRAINING_MIGRATION",
  TRAINING_MAGIC = "TRAINING_MAGIC",
  TRAINING_PHISIC = "TRAINING_PHISIC",

  BACKGROUND_VILLAGE = "BACKGROUND_VILLAGE",
  BACKGROUND_KINGDOM = "BACKGROUND_KINGDOM",
}